Utilizing new env logic from common module + simplified logic
