+++ api/server.go
@@ -16,8 +16,8 @@ import (
 	"github.com/ip2location/ip2location-go/v9"
 	"github.com/rs/cors"
 	cModel "github.com/ygo-skc/skc-go/common/model"
+	cUtil "github.com/ygo-skc/skc-go/common/util"
 	"github.com/ygo-skc/skc-suggestion-engine/db"
-	"github.com/ygo-skc/skc-suggestion-engine/util"
 )
 
 const (
@@ -107,7 +107,7 @@ func commonResponseMiddleware(next http.Handler) http.Handler {
 // Configures routes and their middle wares
 // This method should be called before the environment is set up as the API Key will be set according to the value found in environment
 func RunHttpServer() {
-	serverAPIKey = util.EnvMap["API_KEY"] // configure API Key
+	serverAPIKey = cUtil.EnvMap["API_KEY"] // configure API Key
 	router := mux.NewRouter()
 
 	// configure non-admin routes
