+++ go.mod
@@ -12,7 +12,7 @@ require (
 	github.com/joho/godotenv v1.5.1
 	github.com/rs/cors v1.11.1
 	github.com/stretchr/testify v1.10.0
-	github.com/ygo-skc/skc-go/common v0.0.0-20250124223519-e7b298aaeece
+	github.com/ygo-skc/skc-go/common v0.0.0-20250125014508-6b36cc993f58
 	go.mongodb.org/mongo-driver/v2 v2.0.0
 )
