+++ main.go
@@ -1,10 +1,25 @@
 package main
 
 import (
+	"os"
+	"strings"
+
 	_ "github.com/go-sql-driver/mysql"
+	cUtil "github.com/ygo-skc/skc-go/common/util"
 	"github.com/ygo-skc/skc-suggestion-engine/api"
 	"github.com/ygo-skc/skc-suggestion-engine/db"
 )
+
+const (
+	ENV_VARIABLE_NAME string = "SKC_SUGGESTION_ENGINE_DOT_ENV_FILE"
+)
+
+func init() {
+	isCICD := os.Getenv("IS_CICD")
+	if isCICD != "true" && !strings.HasSuffix(os.Args[0], ".test") {
+		cUtil.ConfigureEnv(ENV_VARIABLE_NAME)
+	}
+}
 
 func main() {
 	db.EstablishDBConn()
