+++ model/suggestion.go
@@ -51,3 +51,26 @@ type ArchetypalSuggestions struct {
 	UsingText  []cModel.Card `json:"usingText"`
 	Exclusions []cModel.Card `json:"exclusions"`
 }
+
+// looks for a self reference, if a self reference is found it is removed from original slice
+// this method returns true if a self reference is found
+func RemoveSelfReference(self string, cr *[]CardReference) bool {
+	hasSelfRef := false
+
+	if cr != nil {
+		x := 0
+		for _, ref := range *cr {
+			if ref.Card.CardName != self {
+				(*cr)[x] = ref
+				x++
+			} else {
+				hasSelfRef = true
+			}
+		}
+
+		*cr = (*cr)[:x]
+		return hasSelfRef
+	} else {
+		return hasSelfRef
+	}
+}
